Modify: ProductCategory 함수형 컴포넌트로 Refactoring
