+++ src/Pages/ProductMain/ProductCategory/ProductCategory.js
@@ -1,75 +1,64 @@
-import React, { Component } from "react";
+import React, { useEffect, useState } from "react";
 import MainCategory from "./MainCategory/MainCategory";
 import "./ProductCategory.scss";
+import { SERVER } from "../../../config";
 
-class ProductCategory extends Component {
-  constructor() {
-    super();
-    this.state = {
-      isListOnOff: true,
-      product: [],
-      selectedCategory: [],
-      productCategory: [],
-    };
-  }
+const ProductCategory = () => {
+  const [productCategory, setProductCategory] = useState([]);
+  const [selectedCategory, setSelectedCategory] = useState([]);
+  const [isListOnOff, setIsListOnOff] = useState(true);
 
-  categoryListOnOff = () => {
-    this.setState({
-      isListOnOff: !this.state.isListOnOff,
-    });
+  useEffect(() => {
+    onCategoryDataRequest();
+  }, []);
+
+  const onCategoryDataRequest = () => {
+    fetch(`${SERVER}/products/category`)
+      .then((response) => response.json())
+      .then((data) => {
+        setProductCategory(data);
+      });
   };
 
-  onhandleListOnOff = (id) => {
-    const selectCategory = this.props.category.categories?.filter((product) => {
+  const categoryListOnOff = () => {
+    setIsListOnOff(!isListOnOff);
+  };
+
+  const onhandleListOnOff = (id) => {
+    const selectCategory = productCategory.categories?.filter((product) => {
       return product.id === id;
     });
-    this.setState({
-      selectedCategory: selectCategory,
-    });
+    setSelectedCategory(selectCategory);
   };
 
-  onSelectCategory = () => {
-    this.props.onDataRequest();
-  };
-
-  componentDidMount() {
-    // this.onhandleListOnOff(1);
-    console.log("1번 카테고리");
-  }
-
-  render() {
-    const { category } = this.props;
-    const { isListOnOff, selectedCategory } = this.state;
-    return (
-      <div className="productCategory">
-        <div className="mainCategory">
-          <MainCategory
-            selectedCategory={selectedCategory}
-            onSelectCategory={this.onSelectCategory}
-            onhandleListOnOff={this.onhandleListOnOff}
-            isListOnOff={isListOnOff}
-            categoryListOnOff={this.categoryListOnOff}
-          />
-        </div>
-        <ul className="otherCategoryList">
-          <h1>
-            {category.categories &&
-              category.categories.map((product, index) => {
-                return (
-                  <div
-                    className="otherCategory"
-                    key={index}
-                    onClick={() => this.onhandleListOnOff(index + 1)}
-                  >
-                    {product.name}
-                  </div>
-                );
-              })}
-          </h1>
-        </ul>
+  return (
+    <div className="productCategory">
+      <div className="mainCategory">
+        <MainCategory
+          selectedCategory={selectedCategory}
+          isListOnOff={isListOnOff}
+          onhandleListOnOff={onhandleListOnOff}
+          categoryListOnOff={categoryListOnOff}
+        />
       </div>
-    );
-  }
-}
+      <ul className="otherCategoryList">
+        <h1>
+          {productCategory.categories &&
+            productCategory.categories.map((product, index) => {
+              return (
+                <div
+                  className="otherCategory"
+                  key={index}
+                  onClick={() => onhandleListOnOff(index + 1)}
+                >
+                  {product.name}
+                </div>
+              );
+            })}
+        </h1>
+      </ul>
+    </div>
+  );
+};
 
 export default ProductCategory;
